# src/swagger/SwaggerDocumentation.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para o projeto Marketplace de produtos farmacêuticos.",
        "version": "1.0.0",
        "title": "Marketplace API"
    },
    "host": "localhost:3002",
    "basePath": "/api",
...
    "schemes": [
        "http"
    ],
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "paths": {
...
        "/user/signup": {
            "post": {
                "tags": [
                    "Usuário"
                ],
                "summary": "Cadastro de usuário",
                "description": "Cadastra um novo usuário comprador",
                "operationId": "createUser",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Dados do usuário",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Sucesso",
                        "schema": {
                            "$ref": "#/definitions/SucessCreated"
                        }
                    },
                    "400": {
                        "description": "Requisição inválida",
                        "schema": {
                            "$ref": "#/definitions/ErrorBAD_REQUEST"
                        }
                    },
                    "409": {
                        "description": "Conflito",
                        "schema": {
                            "$ref": "#/definitions/ErrorCONFLICT"
                        }
                    },
                    "422": {
                        "description": "Entidade inválida",
                        "schema": {
                            "$ref": "#/definitions/ErrorINVALID_DATA"
                        }
                    }
                }
            }
        },
...
        "/sales": {
            "post": {
                "tags": [
                    "Vendas"
                ],
                "summary": "Registra vendas",
                "description": "Registra uma nova venda",
                "operationId": "createSale",
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Itens da venda",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/NewSaleItem"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Criado"
                    },
                    "400": {
                        "description": "Requisição inválida",
                        "schema": {
                            "$ref": "#/definitions/ErrorBAD_REQUEST"
                        }
                    },
                    "401": {
                        "description": "Não autorizado",
                        "schema": {
                            "$ref": "#/definitions/ErrorUNAUTHORIZED"
                        }
                    },
                    "409": {
                        "description": "Conflito",
                        "schema": {
                            "$ref": "#/definitions/ErrorCONFLICT"
                        }
                    },
                    "422": {
                        "description": "Entidade inválida",
                        "schema": {
                            "$ref": "#/definitions/ErrorINVALID_DATA"
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "Vendas"
                ],
                "summary": "Obtém vendas do usuário",
                "description": "Retorna a lista de vendas do usuário autenticado",
                "operationId": "getAuthenticatedUserSales",
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Sale"
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado",
                        "schema": {
                            "$ref": "#/definitions/ErrorUNAUTHORIZED"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ErrorUNAUTHORIZED": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Usuário não autorizado"
                },
                "code": {
                    "type": "string",
                    "example": "UNAUTHORIZED"
                }
            }
        },
...
        "ErrorINVALID_DATA": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "INVALID_DATA"
                },
                "message": {
                    "type": "string",
                    "example": "Dados inválidos"
                },
                "cause": {
                    "type": "string",
                    "example": "O campo 'email' é obrigatório"
                }
            }
        },
        "ErrorCONFLICT": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Usuário já cadastrado"
                },
                "code": {
                    "type": "string",
                    "example": "CONFLICT"
                }
            }
        },
        "ErrorBAD_REQUEST": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Requisição inválida"
                },
                "code": {
                    "type": "string",
                    "example": "BAD_REQUEST"
                }
            }
        },
...
        "SucessCreated": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Usuário criado com sucesso!"
                },
                "code": {
                    "type": "string",
                    "example": "USER_CREATED"
                }
            }
        },
...
        "CreateUserRequest": {
            "type": "object",
            "properties": {
                "user": {
                    "$ref": "#/definitions/User"
                },
                "addresses": {
                    "type": "array",
                    "items": {
                      "$ref": "#/definitions/Address"
                    }
                  }
            }
        },
...
        "Address": {
            "type": "object",
            "properties": {
                "street": {
                    "type": "string"
                },
                "numberStreet": {
                    "type": "integer"
                },
                "complement": {
                    "type": "string"
                },
                "neighborhood": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "zip": {
                    "type": "string"
                },
                "lat": {
                    "type": "string"
                },
                "long": {
                    "type": "string"
                }
            },
            "required": [
                "street",
                "numberStreet",
                "neighborhood",
                "city",
                "state",
                "zip"
            ]
        },
...
        "Product": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "userId": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "labName": {
                    "type": "string"
                },
                "imageLink": {
                    "type": "string",
                    "format": "url",
                    "example": "http://www.teste.com.br"
                },
                "typeDosage": {
                    "type": "string",
                    "enum": [
                        "mg",
                        "mcg",
                        "g",
                        "mL",
                        "%",
                        "outro"
                    ]
                },
                "dosage": {
                    "type": "string",
                    "example": "10"
                },
                "description": {
                    "type": "string"
                },
                "unitPrice": {
                    "type": "number"
                },
                "typeProduct": {
                    "type": "string"
                },
                "totalStock": {
                    "type": "integer"
                }
            }
        },
...
        "User": {
            "type": "object",
            "properties": {
                "fullName": {
                    "type": "string"
                },
                "cpf": {
                    "type": "string",
                    "minLength": 11,
                    "maxLength": 11,
                    "example": "12345678900"
                },
                "birthDate": {
                    "type": "string",
                    "format": "date",
                    "example": "1990-01-01"
                },
                "email": {
                    "type": "string",
                    "format": "email",
                    "example": "teste@teste.com.br"
                },
                "phone": {
                    "type": "string",
                    "minLength": 11,
                    "maxLength": 11,
                    "example": "11999999999"
                },
                "password": {
                    "type": "string",
                    "minLength": 8,
                    "example": "123Senha@"
                },
                "typeUser": {
                    "type": "string",
                    "default": "BUYER"
                }
            },
            "required": [
                "fullName",
                "cpf",
                "birthDate",
                "email",
                "phone",
                "password"
            ]
        },
...
        "NewSaleItem": {
            "type": "object",
            "properties": {
                "productId": {
                    "type": "integer"
                },
                "amountBuy": {
                    "type": "integer"
                },
                "addressId": {
                    "type": "integer"
                },
                "typePayment": {
                    "type": "string"
                }
            }
        },
        "Sale": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "product": {
                    "$ref": "#/definitions/Product"
                },
                "amountBuy": {
                    "type": "integer"
                },
                "total": {
                    "type": "number"
                },
                "createdAt": {
                    "type": "string",
                    "format": "date-time"
                }
            }
        },
...
    }
}